funciona aplicar condicion taxativa!!

+++ ultimoTp/src/main/java/metodologiaFactory/mayor.java
@@ -9,7 +9,7 @@ public class mayor extends expresion{
 	}
 	
 	@Override
-	public boolean cumpleCondicion(int v1, int v2)
+	public boolean cumpleCondicion(double v1, double v2)
 	{
 		return v1>v2;
 	}
